refactor: split character-parser.service.ts into focused modules (#335)
Split the monolithic character-parser.service.ts (671 lines) into 6 focused
modules under character-parser/, following the same barrel pattern as rpy/.

Module breakdown:
- types.ts — all type/interface definitions
- constants.ts — DEFAULT_EXCLUDED_TAGS, INTERPOLATION_REGEX, RENPY_NAMED_COLORS
- color.ts — normalizeColor, extractColor (standalone)
- name-resolution.ts — resolveNameFromSource, classifyName, buildDetectedCharacter
- matchers.ts — tryNullNameMatch, tryStandardMatch, tryMultiLineStart, parseCharacterLine
- parser.ts — CharacterParserService class + singleton

Also within this refactor:
- Extracted RENPY_NAMED_COLORS from inline normalizeColor to module constant
- Extracted resolveNameFromSource() to eliminate duplicated name-resolution logic
- Clarified misleading comment in isSimpleAssignment guard
- Replaced no-op trySimpleAssignment with boolean isSimpleAssignment guard

character-parser.service.ts is now a thin barrel (15 lines) re-exporting
the public API. All existing imports continue to work.
