+++ apps/backend/src/services/character-parser.service.ts
@@ -1,663 +1,15 @@
 /**
- * Character Parser Service
- *
- * Enhanced character definition parser with multi-pattern support.
- * Handles various Ren'Py character definition patterns including:
- * - Standard: define s = Character("Name", color="#...")
- * - Null name: define n = Character(None, ...)
- * - Unknown: define u = Character("???", ...)
- * - Multi-line: definitions spanning multiple lines
- * - Dynamic names: define ne = Character("[persistent.pl_nickname]", ...)
- * - Variable names: define ne = Character(voice_name, ...) or Character([voice_name], ...)
- * - who_color parameter: who_color="#..." instead of color="#..."
- * - Formatted names: define mystery = Character("{color=#f00}Stranger{/color}")
+ * Barrel re-export for backward compatibility.
+ * All implementation lives in ./character-parser/ modules.
  */
 
-// Default excluded character tags (special Ren'Py characters)
-export const DEFAULT_EXCLUDED_TAGS = ["n", "u", "narrator", "extend"] as const;
-export type DefaultExcludedTag = (typeof DEFAULT_EXCLUDED_TAGS)[number];
+export type {
+  DetectedCharacter,
+  CharacterConflict,
+  CharacterParseResult,
+} from "./character-parser/types.js";
 
-/**
- * Detected character from RPY file parsing
- *
- * Field explanations:
- * - tag: The dialogue tag used in RPY files (e.g., "s" for `s "Hello!"`)
- * - name: The raw name as it appeared in the source Character() call,
- *   preserved for reference and possible future round-tripping. The
- *   current BranchForge RPY export emits `displayName`, not `name`, so
- *   do not rely on `name` for export fidelity yet.
- * - displayName: The human-readable name suggested for use in BranchForge
- *   UI (and emitted by the current RPY export). Tags stripped
- *   (e.g., "{color=...}Stranger{/color}" → "Stranger"), variable names
- *   preserved as-is. Empty when the source is `None`/`""`; callers
- *   should derive a fallback (the tag, or `"(unnamed)"`) for display.
- * - nameType: How the name was specified — drives import-wizard warnings.
- * - color: Hex color for dialogue display
- * - isSpecial: Whether this is a system character (narration, unknown speaker)
- * - sourceFile: Which RPY file this was detected in
- * - confidence: Detection confidence (0-1, lower for variable references)
- */
-export interface DetectedCharacter {
-  tag: string;
-  name: string | null;
-  displayName: string;
-  nameType: CharacterNameType;
-  color: string;
-  isSpecial: boolean; // narration, unknown, etc.
-  sourceFile: string;
-  confidence: number; // 0-1 for fuzzy matches
-}
+export { DEFAULT_EXCLUDED_TAGS } from "./character-parser/constants.js";
+export type { DefaultExcludedTag } from "./character-parser/constants.js";
 
-/**
- * Conflict between detected and existing character
- */
-export interface CharacterConflict {
-  tag: string;
-  detectedName: string | null;
-  existingName: string;
-  detectedColor: string;
-  existingColor: string;
-}
-
-/**
- * Character definition parsing result
- */
-export interface CharacterParseResult {
-  characters: DetectedCharacter[];
-  conflicts: CharacterConflict[];
-  excludedTags: Set<string>;
-}
-
-/**
- * Internal form discriminator for how the name was specified in the source.
- * - `quoted` — `Character("Sarah", ...)` or `Character("???", ...)`
- * - `bracketed` — `Character([e_name], ...)` (with square brackets)
- * - `identifier` — `Character(boss_name, ...)` (bare Python identifier)
- */
-type NameForm = "quoted" | "bracketed" | "identifier";
-
-/**
- * Pattern match result for character definitions
- */
-interface CharacterPatternMatch {
-  tag: string;
-  name: string | null;
-  /**
-   * The raw name as written in the source, including any surrounding brackets
-   * (for the `bracketed` form) but excluding the quotes (for the `quoted`
-   * form). Preserved verbatim for round-tripping.
-   */
-  rawName: string | null;
-  nameForm: NameForm | null;
-  /**
-   * True when the name has been fully resolved (set to a value or
-   * explicitly to `null` for `None`). For multi-line starts, the
-   * `parseFile` consumer only continues looking for the name on
-   * subsequent lines when this is `false`. Defaults to `true` for
-   * single-line definitions (which never span multiple lines) and
-   * `false` for multi-line starts whose name will appear later.
-   */
-  nameResolved: boolean;
-  color?: string;
-  isMultiLine: boolean;
-}
-
-import type { CharacterNameType } from "@branchforge/shared";
-import { stripRenpyTextTags } from "@branchforge/shared";
-
-/**
- * Classify a raw extracted name and compute the suggested display name.
- *
- * - `quoted` form:
- *   - `""` → `empty`
- *   - `"???"` → `unknown` (kept verbatim — intentional author choice)
- *   - contains Ren'Py inline tags → `tagged` (displayName = tags stripped)
- *   - contains a `[identifier]` interpolation expression → `interpolated`
- *     (low confidence; Ren'Py resolves the value at runtime)
- *   - otherwise → `literal`
- * - `bracketed` form (`[e_name]`) → `interpolated` (low confidence)
- * - `identifier` form (`boss_name`) → `variable` (low confidence)
- *
- * Known limitation: `INTERPOLATION_REGEX` cannot tell the difference
- * between a Ren'Py variable interpolation (`"[player_name]"`) and a
- * decorative bracket pair in a literal string (`"[END]"`). Both are
- * classified as `interpolated`. The wizard's badge is informational;
- * authors can override the display name on import.
- */
-const INTERPOLATION_REGEX =
-  /\[[a-zA-Z_][a-zA-Z0-9_]*(?:\.[a-zA-Z_][a-zA-Z0-9_]*)*\]/;
-
-function classifyName(
-  rawName: string | null,
-  form: NameForm | null
-): { displayName: string; nameType: CharacterNameType; confidence: number } {
-  if (form === null || rawName === null) {
-    return { displayName: "", nameType: "none", confidence: 0.5 };
-  }
-
-  if (form === "quoted") {
-    if (rawName === "") {
-      return { displayName: "", nameType: "empty", confidence: 1.0 };
-    }
-    if (rawName === "???") {
-      return { displayName: "???", nameType: "unknown", confidence: 1.0 };
-    }
-    if (/\{[^{}]*\}/.test(rawName)) {
-      return {
-        displayName: stripRenpyTextTags(rawName),
-        nameType: "tagged",
-        confidence: 1.0,
-      };
-    }
-    if (INTERPOLATION_REGEX.test(rawName)) {
-      return {
-        displayName: rawName,
-        nameType: "interpolated",
-        confidence: 0.5,
-      };
-    }
-    return { displayName: rawName, nameType: "literal", confidence: 1.0 };
-  }
-
-  if (form === "bracketed") {
-    return { displayName: rawName, nameType: "interpolated", confidence: 0.5 };
-  }
-
-  // form === "identifier"
-  return { displayName: rawName, nameType: "variable", confidence: 0.5 };
-}
-
-/**
- * Build a DetectedCharacter from a pattern match, deriving nameType and
- * displayName via `classifyName`.
- */
-function buildDetectedCharacter(
-  match: CharacterPatternMatch,
-  filename: string,
-  isSpecial: boolean
-): DetectedCharacter {
-  const { displayName, nameType, confidence } = classifyName(
-    match.rawName,
-    match.nameForm
-  );
-  return {
-    tag: match.tag,
-    name: match.name,
-    displayName,
-    nameType,
-    color: match.color || "#cfcfcf",
-    isSpecial,
-    sourceFile: filename,
-    confidence,
-  };
-}
-
-/**
- * Character Parser Service
- */
-class CharacterParserService {
-  /**
-   * Check if a character tag is special (narration, unknown, etc.)
-   */
-  private isSpecialTag(tag: string): boolean {
-    return DEFAULT_EXCLUDED_TAGS.includes(tag as DefaultExcludedTag);
-  }
-
-  /**
-   * Normalize color string to hex format
-   */
-  private normalizeColor(color: string | undefined): string {
-    if (!color) return "#cfcfcf"; // Default Ren'Py color
-
-    // If already hex, return as-is
-    if (color.startsWith("#")) {
-      return color;
-    }
-
-    // Handle named colors (common Ren'Py colors)
-    const namedColors: Record<string, string> = {
-      white: "#ffffff",
-      black: "#000000",
-      red: "#ff0000",
-      green: "#00ff00",
-      blue: "#0000ff",
-      yellow: "#ffff00",
-      cyan: "#00ffff",
-      magenta: "#ff00ff",
-    };
-
-    const normalized = namedColors[color.toLowerCase()];
-    if (normalized) return normalized;
-
-    // Try to extract hex from color string
-    const hexMatch = color.match(/#[0-9a-fA-F]{6}/);
-    if (hexMatch) return hexMatch[0];
-
-    // Default color
-    return "#cfcfcf";
-  }
-
-  /**
-   * Extract color (who_color first, then color) from an options string.
-   */
-  private extractColor(options: string | undefined): string | undefined {
-    if (!options) return undefined;
-    const whoColorMatch = options.match(/who_color\s*=\s*["']?([^"')\s]+)/);
-    if (whoColorMatch) return whoColorMatch[1];
-    const colorMatch = options.match(/color\s*=\s*["']?([^"')\s]+)/);
-    if (colorMatch) return colorMatch[1];
-    return undefined;
-  }
-
-  /**
-   * Parse a single character definition line
-   */
-  private parseCharacterLine(line: string): CharacterPatternMatch | null {
-    const trimmed = line.trim();
-
-    // Skip comments and empty lines
-    if (!trimmed || trimmed.startsWith("#")) {
-      return null;
-    }
-
-    // Skip simple variable assignments (e.g., define ne_first = "Lucas")
-    // These are not character definitions, just setting up variables
-    const simpleAssignmentMatch = trimmed.match(
-      /(?:define|default)\s+([a-zA-Z_][a-zA-Z0-9_]*)\s*=\s*"([^"]+)"/
-    );
-    if (simpleAssignmentMatch) {
-      // Check if this is followed by " = Character" on next lines (not relevant for single-line parsing)
-      // For now, skip these - they're just variable assignments
-      return null;
-    }
-
-    // Null name pattern: define n = Character(None, options...)
-    // Also supports: default n = Character(None, options...)
-    // Check this BEFORE the standard pattern because `None` would otherwise
-    // match the identifier group of the standard pattern.
-    const nullNameMatch = trimmed.match(
-      /(?:define|default)\s+([a-zA-Z_][a-zA-Z0-9_]*)\s*=\s*Character\s*\(\s*None(\s*,\s*([^)]*))?\s*\)/
-    );
-    if (nullNameMatch) {
-      const tag = nullNameMatch[1];
-      const options = nullNameMatch[2];
-
-      const color = this.extractColor(options);
-
-      return {
-        tag,
-        name: null,
-        rawName: null,
-        nameForm: null,
-        nameResolved: true,
-        color: this.normalizeColor(color),
-        isMultiLine: false,
-      };
-    }
-
-    // Standard single-line pattern: define tag = Character("name", options...)
-    // Also supports: default tag = Character("name", options...)
-    // Also supports variable names: Character(variable_name, ...) or Character([variable_name], ...)
-    const standardMatch = trimmed.match(
-      /(?:define|default)\s+([a-zA-Z_][a-zA-Z0-9_]*)\s*=\s*Character\s*\(\s*(?:"([^"]*)"|(\[([a-zA-Z_][a-zA-Z0-9_.]*)\])|([a-zA-Z_][a-zA-Z0-9_.]*))(\s*,\s*([^)]*))?\s*\)/
-    );
-    if (standardMatch) {
-      const tag = standardMatch[1];
-      // Groups: 2=quoted, 3=bracketed-full (e.g., "[e_name]"), 4=bracketed-inner,
-      // 5=identifier, 6=options-after-comma, 7=options-content
-      let name: string | null = null;
-      let rawName: string | null = null;
-      let nameForm: NameForm | null = null;
-      if (standardMatch[2] !== undefined) {
-        name = standardMatch[2];
-        rawName = standardMatch[2];
-        nameForm = "quoted";
-      } else if (standardMatch[3] !== undefined) {
-        // Both `name` and `rawName` keep the brackets for round-trip
-        // fidelity. This keeps the bracketed arg form consistent with
-        // the quoted form `"[e_name]"` (where the brackets appear in
-        // the string content).
-        name = standardMatch[3]; // includes brackets, e.g. "[e_name]"
-        rawName = standardMatch[3];
-        nameForm = "bracketed";
-      } else if (standardMatch[5] !== undefined) {
-        rawName = standardMatch[5];
-        name = standardMatch[5];
-        nameForm = "identifier";
-      }
-      const options = standardMatch[7];
-
-      const color = this.extractColor(options);
-
-      return {
-        tag,
-        name,
-        rawName,
-        nameForm,
-        nameResolved: true,
-        color: this.normalizeColor(color),
-        isMultiLine: false,
-      };
-    }
-
-    // Multi-line start pattern: define tag = Character(...)
-    // Also supports: default tag = Character(...)
-    const multiLineStartMatch = trimmed.match(
-      /(?:define|default)\s+([a-zA-Z_][a-zA-Z0-9_]*)\s*=\s*Character\s*\((.*)/
-    );
-    if (multiLineStartMatch) {
-      const tag = multiLineStartMatch[1];
-      const rest = multiLineStartMatch[2];
-
-      // Check if name is on the same line (quoted string, bracketed, or identifier)
-      let name: string | null = null;
-      let rawName: string | null = null;
-      let nameForm: NameForm | null = null;
-      // True once we have a final answer for the name — either a
-      // detected form, or explicit `None`. False when the name will
-      // appear on a later line of the multi-line definition.
-      let nameResolved = false;
-
-      const quotedNameMatch = rest.match(/"([^"]*)"/);
-      if (quotedNameMatch) {
-        name = quotedNameMatch[1];
-        rawName = quotedNameMatch[1];
-        nameForm = "quoted";
-        nameResolved = true;
-      } else if (
-        // Explicit `None` token on the multi-line start line, e.g.
-        // `define n = Character(None,`. The standard `nullNameMatch`
-        // in `parseCharacterLine` only catches single-line
-        // definitions; here we need to handle the multi-line form
-        // before the identifier fallback would otherwise miscapture
-        // `None` as a variable name.
-        /^None\b/.test(rest) ||
-        /^\s*None\s*[,)]/.test(rest)
-      ) {
-        name = null;
-        rawName = null;
-        nameForm = null;
-        nameResolved = true;
-      } else {
-        const bracketedNameMatch = rest.match(/\[([a-zA-Z_][a-zA-Z0-9_.]*)\]/);
-        if (bracketedNameMatch) {
-          // Both `name` and `rawName` keep the brackets for round-trip
-          // fidelity (matches the single-line `Character([e_name], ...)`
-          // case).
-          name = `[${bracketedNameMatch[1]}]`;
-          rawName = `[${bracketedNameMatch[1]}]`;
-          nameForm = "bracketed";
-          nameResolved = true;
-        } else {
-          const variableNameMatch = rest.match(/([a-zA-Z_][a-zA-Z0-9_.]*)/);
-          if (variableNameMatch) {
-            rawName = variableNameMatch[1];
-            name = variableNameMatch[1];
-            nameForm = "identifier";
-            nameResolved = true;
-          }
-        }
-      }
-
-      // Try to extract color from rest of line (who_color first, then color)
-      let color: string | undefined = undefined;
-      if (rest.trim()) {
-        color = this.extractColor(rest);
-      }
-
-      return {
-        tag,
-        name,
-        rawName,
-        nameForm,
-        // Pass through the explicit nameResolved signal (covers
-        // detected form, explicit `None`, and unresolved cases).
-        nameResolved,
-        color: this.normalizeColor(color),
-        isMultiLine: true,
-      };
-    }
-
-    return null;
-  }
-
-  /**
-   * Parse character definitions from a single file
-   */
-  parseFile(content: string, filename: string): DetectedCharacter[] {
-    const characters: DetectedCharacter[] = [];
-    const lines = content.split("\n");
-
-    // Track multi-line definitions
-    let pendingCharacter: {
-      tag: string;
-      name: string | null;
-      rawName: string | null;
-      nameForm: NameForm | null;
-      color: string | undefined;
-      options: string[];
-      startLine: number;
-      /**
-       * True once the name has been resolved (or explicitly set to
-       * null for `None`). Prevents subsequent option lines like
-       * `color="#cfcfcf"` from re-matching as the character name.
-       */
-      nameResolved: boolean;
-    } | null = null;
-    let parenDepth = 0;
-
-    for (let i = 0; i < lines.length; i++) {
-      const line = lines[i];
-      const trimmed = line.trim();
-
-      // Skip empty lines and comments
-      if (!trimmed || trimmed.startsWith("#")) {
-        continue;
-      }
-
-      // Check if we're in a multi-line definition
-      if (pendingCharacter) {
-        // Track parentheses
-        parenDepth += (line.match(/\(/g) || []).length;
-        parenDepth -= (line.match(/\)/g) || []).length;
-
-        // Capture options
-        if (trimmed && !trimmed.startsWith("#")) {
-          // Check if this line contains the name (if not already found)
-          if (!pendingCharacter.nameResolved) {
-            // Try quoted name. Use `*` (not `+`) to also match
-            // empty quoted strings like `""`, which Ren'Py treats as
-            // a literal "no display name".
-            const quotedNameMatch = trimmed.match(/"([^"]*)"/);
-            if (quotedNameMatch) {
-              pendingCharacter.name = quotedNameMatch[1];
-              pendingCharacter.rawName = quotedNameMatch[1];
-              pendingCharacter.nameForm = "quoted";
-              pendingCharacter.nameResolved = true;
-            } else if (
-              /^None\b/.test(trimmed) ||
-              /^\s*None\s*[,)]/.test(trimmed)
-            ) {
-              // Explicit `None` token: the character has no display name.
-              // (Ren'Py treats `Character(None, ...)` as the narrator.)
-              pendingCharacter.name = null;
-              pendingCharacter.rawName = null;
-              pendingCharacter.nameForm = null;
-              pendingCharacter.nameResolved = true;
-            } else {
-              // Try bracketed name
-              const bracketedNameMatch = trimmed.match(
-                /\[([a-zA-Z_][a-zA-Z0-9_.]*)\]/
-              );
-              if (bracketedNameMatch) {
-                // Both name and rawName keep the brackets for
-                // round-trip fidelity.
-                pendingCharacter.rawName = `[${bracketedNameMatch[1]}]`;
-                pendingCharacter.name = `[${bracketedNameMatch[1]}]`;
-                pendingCharacter.nameForm = "bracketed";
-                pendingCharacter.nameResolved = true;
-              } else {
-                // Try variable name
-                const variableNameMatch = trimmed.match(
-                  /([a-zA-Z_][a-zA-Z0-9_.]*)/
-                );
-                if (variableNameMatch) {
-                  pendingCharacter.rawName = variableNameMatch[1];
-                  pendingCharacter.name = variableNameMatch[1];
-                  pendingCharacter.nameForm = "identifier";
-                  pendingCharacter.nameResolved = true;
-                }
-              }
-            }
-          }
-          pendingCharacter.options.push(trimmed);
-        }
-
-        // Check if definition is complete
-        if (parenDepth <= 0) {
-          // Extract color from options (who_color first, then color)
-          let color = pendingCharacter.color;
-          if (!color) {
-            const optionsText = pendingCharacter.options.join(" ");
-            color = this.extractColor(optionsText);
-          }
-
-          const isSpecial = this.isSpecialTag(pendingCharacter.tag);
-          const match: CharacterPatternMatch = {
-            tag: pendingCharacter.tag,
-            name: pendingCharacter.name,
-            rawName: pendingCharacter.rawName,
-            nameForm: pendingCharacter.nameForm,
-            // The continuation loop only runs while the name is
-            // unresolved, so by the time we get here the name has
-            // been resolved one way or another.
-            nameResolved: true,
-            color: this.normalizeColor(color),
-            isMultiLine: true,
-          };
-          characters.push(buildDetectedCharacter(match, filename, isSpecial));
-
-          pendingCharacter = null;
-          parenDepth = 0;
-        }
-        continue;
-      }
-
-      // Try to parse as a character definition
-      const match = this.parseCharacterLine(line);
-      if (match) {
-        if (match.isMultiLine) {
-          // Start tracking multi-line definition
-          parenDepth =
-            (line.match(/\(/g) || []).length - (line.match(/\)/g) || []).length;
-          pendingCharacter = {
-            tag: match.tag,
-            name: match.name,
-            rawName: match.rawName,
-            nameForm: match.nameForm,
-            color: match.color,
-            options: [],
-            startLine: i,
-            // Pass through the parser's name-resolved signal. When
-            // false, the continuation loop will look for the name
-            // on subsequent lines.
-            nameResolved: match.nameResolved,
-          };
-        } else {
-          // Single-line definition
-          const isSpecial = this.isSpecialTag(match.tag);
-          characters.push(buildDetectedCharacter(match, filename, isSpecial));
-        }
-      }
-    }
-
-    return characters;
-  }
-
-  /**
-   * Parse with exclusions applied
-   */
-  parseWithExclusions(
-    content: string,
-    filename: string,
-    excludedTags: Set<string>
-  ): DetectedCharacter[] {
-    const allCharacters = this.parseFile(content, filename);
-    return allCharacters.filter((c) => !excludedTags.has(c.tag));
-  }
-
-  /**
-   * Detect conflicts between detected and existing characters
-   */
-  detectConflicts(
-    detected: DetectedCharacter[],
-    existing: Array<{
-      renpyTag: string;
-      name: string;
-      displayName: string;
-      color: string;
-    }>
-  ): CharacterConflict[] {
-    const conflicts: CharacterConflict[] = [];
-    const existingByTag = new Map(existing.map((c) => [c.renpyTag, c]));
-
-    for (const detectedChar of detected) {
-      const existingChar = existingByTag.get(detectedChar.tag);
-
-      if (existingChar) {
-        // Check for differences
-        const nameMismatch = detectedChar.name !== existingChar.name;
-        const displayNameMismatch =
-          detectedChar.displayName !== existingChar.displayName;
-        const colorMismatch = detectedChar.color !== existingChar.color;
-
-        if (nameMismatch || displayNameMismatch || colorMismatch) {
-          conflicts.push({
-            tag: detectedChar.tag,
-            detectedName: detectedChar.name,
-            existingName: existingChar.name,
-            detectedColor: detectedChar.color,
-            existingColor: existingChar.color,
-          });
-        }
-      }
-    }
-
-    return conflicts;
-  }
-
-  /**
-   * Parse multiple files and aggregate results
-   */
-  parseFiles(
-    files: Array<{ content: string; filename: string }>
-  ): CharacterParseResult {
-    const allCharacters: DetectedCharacter[] = [];
-    const excludedTags = new Set<string>(DEFAULT_EXCLUDED_TAGS);
-
-    for (const file of files) {
-      const characters = this.parseFile(file.content, file.filename);
-      allCharacters.push(...characters);
-    }
-
-    // Deduplicate by tag (keep first occurrence)
-    const seenTags = new Set<string>();
-    const uniqueCharacters: DetectedCharacter[] = [];
-
-    for (const char of allCharacters) {
-      if (!seenTags.has(char.tag)) {
-        seenTags.add(char.tag);
-        uniqueCharacters.push(char);
-      }
-    }
-
-    return {
-      characters: uniqueCharacters,
-      conflicts: [],
-      excludedTags,
-    };
-  }
-}
-
-// Export singleton instance
-export const characterParserService = new CharacterParserService();
+export { characterParserService } from "./character-parser/parser.js";
